requirements.txt
pydantic==1.10.7
pandas==2.0.0
tika==2.6.0
lxml==4.9.2
langdetect==1.0.9
XlsxWriter==3.0.9
openpyxl==3.1.2
spacy==3.5.3
wget==3.2
sentence-transformers==2.2.2
harmonydata==0.5.2
openai==0.27.8